+++ Sources/ATTNConstants.m
@@ -7,6 +7,8 @@
 
 #import "ATTNConstants.h"
 
+// MARK: ATTNSDK
+
 // Status passed to ATTNCreativeTriggerCompletionHandler when the creative is opened sucessfully
 NSString *const CREATIVE_TRIGGER_STATUS_OPENED = @"CREATIVE_TRIGGER_STATUS_OPENED";
 // Status passed to ATTNCreativeTriggerCompletionHandler when the creative is closed sucessfully
@@ -18,3 +20,18 @@ NSString *const CREATIVE_TRIGGER_STATUS_NOT_OPENED = @"CREATIVE_TRIGGER_STATUS_N
 // Status passed to the ATTNCreativeTriggerCompletionHandler when the Creative is not closed due to an unknown
 // exception
 NSString *const CREATIVE_TRIGGER_STATUS_NOT_CLOSED = @"CREATIVE_TRIGGER_STATUS_NOT_CLOSED";
+
+// MARK: ATTNUserIdentify
+
+// Your unique identifier for the user - this should be consistent across the user's lifetime, for example a database id
+NSString *const IDENTIFIER_TYPE_CLIENT_USER_ID = @"clientUserId";
+// The user's phone number in E.164 format
+NSString *const IDENTIFIER_TYPE_PHONE = @"phone";
+// The user's email
+NSString *const IDENTIFIER_TYPE_EMAIL = @"email";
+// The user's Shopify Customer ID
+NSString *const IDENTIFIER_TYPE_SHOPIFY_ID = @"shopifyId";
+// The user's Klaviyo ID
+NSString *const IDENTIFIER_TYPE_KLAVIYO_ID = @"klaviyoId";
+// Key-value pairs of custom identifier names and values (both NSStrings) to associate with this user
+NSString *const IDENTIFIER_TYPE_CUSTOM_IDENTIFIERS = @"customIdentifiers";
