+++ Sources/ATTNUserIdentity.m
@@ -10,21 +10,7 @@
 #import "ATTNParameterValidation.h"
 #import "ATTNUserIdentity.h"
 #import "ATTNVisitorService.h"
-
-
-// Your unique identifier for the user - this should be consistent across the user's lifetime, for example a database id
-NSString *const IDENTIFIER_TYPE_CLIENT_USER_ID = @"clientUserId";
-// The user's phone number in E.164 format
-NSString *const IDENTIFIER_TYPE_PHONE = @"phone";
-// The user's email
-NSString *const IDENTIFIER_TYPE_EMAIL = @"email";
-// The user's Shopify Customer ID
-NSString *const IDENTIFIER_TYPE_SHOPIFY_ID = @"shopifyId";
-// The user's Klaviyo ID
-NSString *const IDENTIFIER_TYPE_KLAVIYO_ID = @"klaviyoId";
-// Key-value pairs of custom identifier names and values (both NSStrings) to associate with this user
-NSString *const IDENTIFIER_TYPE_CUSTOM_IDENTIFIERS = @"customIdentifiers";
-
+#import "ATTNConstants.h"
 
 @implementation ATTNUserIdentity {
   ATTNVisitorService *_visitorService;
