+++ Sources/Migration/ATTNSDKMode.swift
@@ -1,0 +1,14 @@
+//
+//  ATTNSDKMode.swift
+//  attentive-ios-sdk-framework
+//
+//  Created by Vladimir - Work on 2024-05-29.
+//
+
+import Foundation
+
+/// Define initialization mode for SDK
+public enum ATTNSDKMode: String {
+  case debug
+  case production
+}
