+++ Sources/include/ATTNUserIdentity.h
@@ -12,14 +12,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-extern NSString * const IDENTIFIER_TYPE_CLIENT_USER_ID;
-extern NSString * const IDENTIFIER_TYPE_PHONE;
-extern NSString * const IDENTIFIER_TYPE_EMAIL;
-extern NSString * const IDENTIFIER_TYPE_SHOPIFY_ID;
-extern NSString * const IDENTIFIER_TYPE_KLAVIYO_ID;
-extern NSString * const IDENTIFIER_TYPE_CUSTOM_IDENTIFIERS;
-
-
 @interface ATTNUserIdentity : NSObject
 
 @property NSDictionary * identifiers;
